nimble/ll: Add CS Configuration procedure
Implements:
- LE CS Create Config command
- LE CS Remove Config command

+++ nimble/controller/include/controller/ble_ll_conn.h
@@ -279,7 +279,7 @@ struct ble_ll_conn_sm
     uint8_t vers_nr;
     uint8_t conn_features;
     uint8_t remote_features[7];
-    uint16_t pending_ctrl_procs;
+    uint32_t pending_ctrl_procs;
     uint16_t event_cntr;
     uint16_t completed_pkts;
     uint16_t comp_id;
